+++ jgclark.Dashboard/src/react/components/ItemNoteLink.jsx
@@ -10,7 +10,7 @@ import type { TSection, TSectionItem } from '../../types.js'
 import { useAppContext } from './AppContext.jsx'
 import TooltipOnKeyPress from './ToolTipOnModifierPress.jsx'
 import NoteTitleLink from './NoteTitleLink.jsx'
-import { logDebug, logInfo } from '@helpers/react/reactDev.js'
+import { logDebug, logInfo } from '@helpers/dev'
 import { parseTeamspaceFilename, TEAMSPACE_FA_ICON } from '@helpers/teamspace'
 
 //-----------------------------------------------------------