+++ jgclark.Dashboard/src/react/components/TaskItem.jsx
@@ -10,7 +10,7 @@ import React, { type Node, useState } from 'react'
 import type { MessageDataObject, TSection, TSectionItem } from '../../types'
 import ItemContent from './ItemContent.jsx'
 import StatusIcon from './StatusIcon.jsx'
-import { clo, JSP, logDebug, logInfo, logWarn } from '@helpers/react/reactDev.js'
+import { clo, JSP, logDebug, logInfo, logWarn } from '@helpers/dev'
 import './TaskItem.css'
 
 type Props = {